refactored test runner to allow for being run as esm
