+++ Makefile
@@ -11,11 +11,14 @@ test-chrome:
 test-firefox:
 	karma start --single-run --browsers=Firefox
 
+
 test-unit:
 	mocha test/server.js
 
 test-node6: transpile
 	node test/node6.js
+
+test: test-unit
 
 typecheck:
 	dtslint --expectOnly types
